move bots into Game struct
